requirements.txt
streamlit
PyPDF2
pdfplumber
python-docx
PyMuPDF==1.23.9
requests
